Fortschritt meines Abschlussprojektes (Stand: 01.02. 15:56 Uhr) -> weitere Funktionen erstellt, -> Schleifen erweitert, -> Schleifen erneuert

+++ src/main/kotlin/Gegner.kt
@@ -1,4 +1,4 @@
-open class Gegner(var name: String, var hp: Int, var schaden: Int) {
+open class Gegner(var name: String, var hp: Double, var schaden: Int) {
 
     var faehigkeiten = mutableListOf<String>(
         "Klon heraufbeschwören",
